separate channeltrackingstatus
